+++ groovy/material/SuSyMaterials.groovy
@@ -1634,8 +1634,7 @@ class SuSyMaterials {
     public static Material YttriaStabilizedZirconia
     public static Material BismuthIIINitrateSolution
     public static Material LeadIVOxide
-    public static Material AluminiumAlloy6061
-    public static Material AluminiumAlloy7075
+    public static Material ManganeseSteel
 
     // HP Materials
