+++ groovy/postInit/mod/MachineRecipes.groovy
@@ -1514,3 +1514,111 @@ RecyclingHelper.handleRecycling(metaitem('susy:gas_turbine'), [
     ore('wireGtQuadrupleAluminium') * 4,
     ore('ringSilicone') * 8,
 ])
+
+// Crushers and grinders
+
+ASSEMBLER.recipeBuilder()
+    .circuitMeta(11)
+    .inputs(metaitem('hull.mv'))
+    .inputs(metaitem('electric.motor.mv') * 4)
+    .inputs(metaitem('gearAluminium') * 2)
+    .inputs(metaitem('conveyor.module.mv') * 2)
+    .inputs(ore('circuitMv') * 4)
+    .fluidInputs(fluid('lubricant') * 4000)
+    .outputs(metaitem('susy:eccentric_roll_crusher'))
+    .EUt(VA[MV])
+    .duration(600)
+    .buildAndRegister()
+
+ASSEMBLER.recipeBuilder()
+    .circuitMeta(11)
+    .inputs(ore('stickLongSteel'))
+    .inputs(ore('frameSteel'))
+    .inputs(ore('screwSteel') * 32)
+    .inputs(ore('plateDoubleManganeseSteel') * 8)
+    .outputs(item('susy:eccentric_roll'))
+    .EUt(VA[MV])
+    .duration(300)
+    .buildAndRegister()
+
+ASSEMBLER.recipeBuilder()
+    .circuitMeta(11)
+    .inputs(ore('frameSteel'))
+    .inputs(ore('plateManganeseSteel') * 6)
+    .outputs(item('susy:grinder_casing'))
+    .EUt(VA[MV])
+    .duration(80)
+    .buildAndRegister()
+
+ASSEMBLER.recipeBuilder()
+    .circuitMeta(11)
+    .inputs(ore('frameSteel'))
+    .inputs(ore('plateSteel') * 6)
+    .inputs(ore('rotorSteel') * 2)
+    .inputs(ore('gearSmallSteel') * 4)
+    .inputs(ore('stickLongSteel') * 2)
+    .inputs(ore('ringRubber') * 4)
+    .fluidInputs(fluid('lubricant') * 4000)
+    .outputs(item('susy:grinder_casing', 1))
+    .EUt(VA[MV])
+    .duration(80)
+    .buildAndRegister()
+
+ASSEMBLER.recipeBuilder()
+    .circuitMeta(12)
+    .inputs(metaitem('hull.mv'))
+    .inputs(metaitem('electric.motor.mv') * 4)
+    .inputs(metaitem('gearAluminium') * 2)
+    .inputs(metaitem('conveyor.module.mv') * 2)
+    .inputs(ore('circuitMv') * 4)
+    .outputs(metaitem('susy:ball_mill'))
+    .EUt(VA[MV])
+    .duration(600)
+    .buildAndRegister()
+
+EXTRUDER.recipeBuilder()
+    .circuitMeta(12)
+    .inputs(ore('ingotSteel') * 9)
+    .outputs(item('susy:girth_gear_tooth'))
+    .EUt(VA[MV])
+    .duration(600)
+    .buildAndRegister()
+
+ASSEMBLER.recipeBuilder()
+    .circuitMeta(12)
+    .inputs(ore('frameSteel'))
+    .inputs(ore('plateSteel') * 6)
+    .inputs(ore('plateManganeseSteel'))
+    .inputs(ore('screwSteel') * 4)
+    .outputs(item('susy:grinder_casing', 2))
+    .EUt(VA[MV])
+    .duration(80)
+    .buildAndRegister()
+
+ASSEMBLER.recipeBuilder()
+    .circuitMeta(12)
+    .outputs(item('susy:grinder_casing', 3))
+    .inputs(ore('frameSteel'))
+    .inputs(ore('plateDoubleSteel') * 6)
+    .inputs(ore('plateDoubleManganeseSteel'))
+    .inputs(ore('screwSteel') * 4)
+    .outputs(item('susy:grinder_casing', 2))
+    .EUt(VA[MV])
+    .duration(80)
+    .buildAndRegister()
+
+BENDER.recipeBuilder()
+    .circuitMeta(1)
+    .inputs(ore('frameSteel'))
+    .outputs(item('susy:grinder_casing', 4))
+    .EUt(VA[MV])
+    .duration(100)
+    .buildAndRegister()
+
+FLUID_SOLIDIFIER.recipeBuilder()
+    .notConsumable(metaitem('shape.mold.ball'))
+    .fluidInputs(fluid('steel') * 144)
+    .outputs(metaitem('millBallSteel'))
+    .EUt(VA[LV])
+    .duration(200)
+    .buildAndRegister()
