+++ resources/langfiles/lang/en_us.lang
@@ -2774,6 +2774,7 @@ susy.material.pitch=Pitch
 susy.material.tabular_alumina=Tabular Alumina
 susy.material.reactive_alumina=Reactive Alumina
 susy.material.tabular_alumina_refractory=Tabular Alumina Refractory
+susy.material.manganese_steel=Manganese Steel
 
 # Isotopes
